requirements.txt
# Core AI packages
python-dotenv    # load environment variables from .env
fastapi
uvicorn[standard]
pydantic>=2
sentence-transformers
faiss-cpu
tqdm

# Data science stack
numpy

# Dev tools
black            # code formatter
ruff             # linter / static analysis


